Add commented out alias overrides
Eventually, worldedit *will* become an optional dependency. The
rationale for this is that WorldEditAdditions is outgrowing the core
WorldEdit API, and we want to add new features such as toggling
safe_region on and off with a chat command and other such goodies.

Merging the 2 mods is not something that has been discussed (due
mainly because I'm far too nervous to even ask the question in the
first place), but the 2 codebases are fundamentally different (and
for good reason, as WorldEditAdditions splits code over many different
files to improve maintainability and scalability) so this would
be a significant undertaking.

At no point however will WorldEditAdditions become incompatible
with WorldEdit itself. The 2 mods should happily co-exist with
one another (so long as you keep them both updated of course).